panics: move PanicCatcher to separate package
